Separate blueprint validation code
