+++ pipelex/create/build_blueprint.py
@@ -1,25 +1,18 @@
 from __future__ import annotations
 
-import asyncio
 from pathlib import Path
-from typing import Any, Dict, Optional, cast
+from typing import Optional
 
 import typer
-from pydantic import ValidationError
 
-from pipelex import log, pretty_print
+from pipelex import pretty_print
 from pipelex.create.helpers import get_support_file
 from pipelex.create.pipeline_toml import save_pipeline_blueprint_toml_to_path
-from pipelex.exceptions import PipeDefinitionError, PipelexCLIError
-from pipelex.hub import get_library_manager
-from pipelex.libraries.library_manager import LibraryManager
+from pipelex.create.validate_blueprint import validate_blueprint
 from pipelex.libraries.pipeline_blueprint import PipelineBlueprint
-from pipelex.pipe_works.pipe_dry import dry_run_pipe_codes
 from pipelex.pipeline.execute import execute_pipeline
 from pipelex.tools.misc.file_utils import save_text_to_path
 from pipelex.tools.misc.json_utils import save_as_json_to_path
-from pipelex.tools.misc.toml_utils import load_toml_from_path
-from pipelex.tools.typing.pydantic_utils import format_pydantic_validation_error
 
 
 async def do_build_blueprint(
@@ -56,95 +49,4 @@ async def do_build_blueprint(
     typer.echo(f"✅ Blueprint saved to '{output_path}'")
 
     if validate:
-        try:
-            _load_pipes_from_generated_blueprint(blueprint=blueprint)
-        except PipeDefinitionError as exc:
-            raise PipelexCLIError(f"Failed to load pipes from generated blueprint at '{output_path}': {exc}") from exc
-
-        generated_pipe_codes = list(blueprint.pipe.keys())
-        if not generated_pipe_codes:
-            raise PipelexCLIError("No pipe found in generated blueprint to validate")
-        log.info(f"Loaded pipes: {generated_pipe_codes}")
-
-        log.info("Validating pipes...")
-        asyncio.run(dry_run_pipe_codes(pipe_codes=generated_pipe_codes))
-        log.info(f"Pipes validated: {generated_pipe_codes}")
-
-
-def _load_pipes_from_generated_blueprint(blueprint: PipelineBlueprint) -> None:
-    """Instantiate and register all pipes from a generated PipelineBlueprint.
-
-    This constructs the appropriate typed PipeBlueprint for each pipe entry and
-    uses LibraryManager.load_pipe_from_blueprint to create the Pipe, then adds it
-    to the in-memory PipeLibrary so that validation can run against them.
-    """
-    library_manager = cast(LibraryManager, get_library_manager())
-    domain_code = blueprint.domain
-
-    for pipe_code, details in blueprint.pipe.items():
-        details_dict: Dict[str, Any] = details.copy()
-
-        # Build the concrete PipeBlueprint via LibraryManager helper
-        # try:
-        #     pipe_blueprint = LibraryManager.make_pipe_blueprint_from_details(
-        #         domain_code=domain_code,
-        #         details_dict=details_dict,
-        #     )
-        # except ValidationError as exc:
-        #     error_msg = format_pydantic_validation_error(exc=exc)
-        #     raise PipeDefinitionError(f"Failed to build pipe blueprint for pipe '{pipe_code}': {error_msg}\n{exc}") from exc
-
-        pipe_blueprint = LibraryManager.make_pipe_blueprint_from_details(
-            domain_code=domain_code,
-            details_dict=details_dict,
-        )
-        # Create pipe and register in library
-        pipe = LibraryManager.load_pipe_from_blueprint(
-            pipe_code=pipe_code,
-            pipe_blueprint=pipe_blueprint,
-        )
-        library_manager.pipe_library.add_new_pipe(pipe=pipe)
-
-
-async def validate_blueprint(blueprint_path: str) -> None:
-    """Validate an already generated pipeline blueprint from a TOML file.
-
-    Args:
-        blueprint_path: Path to the blueprint TOML file to validate
-
-    Raises:
-        PipelexCLIError: If validation fails
-        ValidationError: If the TOML structure doesn't match PipelineBlueprint
-    """
-    # Load the TOML file
-    log.info(f"Loading blueprint from '{blueprint_path}'...")
-    blueprint_dict = load_toml_from_path(path=blueprint_path)
-
-    # Validate and create PipelineBlueprint model
-    try:
-        blueprint = PipelineBlueprint.model_validate(blueprint_dict)
-    except ValidationError as exc:
-        error_msg = format_pydantic_validation_error(exc=exc)
-        raise PipelexCLIError(f"Invalid blueprint structure in '{blueprint_path}': {error_msg}") from exc
-
-    log.info(f"Blueprint loaded successfully: domain='{blueprint.domain}'")
-
-    # Load pipes from the blueprint
-    try:
-        _load_pipes_from_generated_blueprint(blueprint=blueprint)
-    except PipeDefinitionError as exc:
-        raise PipelexCLIError(f"Failed to load pipes from blueprint at '{blueprint_path}': {exc}") from exc
-
-    # Get pipe codes and validate them
-    generated_pipe_codes = list(blueprint.pipe.keys())
-    if not generated_pipe_codes:
-        raise PipelexCLIError("No pipe found in blueprint to validate")
-    log.info(f"Loaded pipes: {generated_pipe_codes}")
-
-    # Run validation
-    log.info("Validating pipes ----------------------")
-    log.info("Validating libraries ----------------------")
-    get_library_manager().validate_libraries()
-    log.info("Dry running pipes...")
-    await dry_run_pipe_codes(pipe_codes=generated_pipe_codes)
-    log.info(f"✅ All pipes validated successfully: {generated_pipe_codes}")
+        await validate_blueprint(blueprint=blueprint)
